Add expiring trash objects
Set "trash_lifetime = N" in the middleware's config to keep trash
objects for only N seconds. Default is 90 days.

This would change behavior for existing deployments if there were any,
but the project is all of 4 days old and is at version 0.0.1, so the
probability of there being any existing deployments is essentially 0.