Allow for blocking of DELETEs to trash objects
If you set "block_trash_deletes = yes" in the config, objects in trash
containers will not be deletable. Clients will get a 405 Method Not
Allowed in response to a DELETE of a trash object.

Note that this commit does not modify the OPTIONS response to remove
DELETE from the Allow header. That should probably happen at some
point.